Initial File Dump
Contains basic game structure and functionality. Game is entirely playable just devoid of much content and maps which can be easily added by anyone with a little python knowledge. I'll personally be adding a fair degree more to this over time but this is just the initial few hours of work I put into this idea.